Desktop/route_fuel_api_full/requirements.txt
Django>=4.2
djangorestframework>=3.15
requests>=2.31
polyline>=1.5.0
django-cors-headers>=4.0
